# parser: Improve TOML frontmatter parser performance